Split pool tests
