+++ tests/pools/general/test_virtual_price.py
